New Foot IK!!!
Boy this was a brainscratcher
Major AHA moment  was figuring out that the rig objects need actual bones to have "parent power". ???
Couldn't just use a gameobject as a parent to other objects that constrain bones.
(the oarent object controls the gameobjects, but the bones were unaffected.
It needed to actually be connected to a frikkin BONE in order to have a parental effect over the bones that its CHILDREN were constraining. How weird.
Guess thats just how Unity's Animation RIgging System works ?